+++ flutter_angle/lib/desktop/wrapper.dart
@@ -138,7 +138,7 @@ class RenderingContext {
     var indices = Pointer<Void>.fromAddress(offset);
     gl.glDrawElementsInstanced(mode, count, type, indices, instanceCount);
     checkError('drawElementsInstanced');
-    calloc.free(indices);
+    //calloc.free(indices);
   }
 
   void endTransformFeedback(){
@@ -147,7 +147,7 @@ class RenderingContext {
     checkError('endTransformFeedback');
   }
 
-  ActiveInfo getTransformFeedbackVarying(int program, int index) {
+  ActiveInfo getTransformFeedbackVarying(Program program, int index) {
     startCheck('getTransformFeedbackVarying');
     int maxLen = 100;
     var length = calloc<Int32>();
@@ -155,7 +155,7 @@ class RenderingContext {
     var type = calloc<Uint32>();
     var name = calloc<Int8>(maxLen);
 
-    gl.glGetTransformFeedbackVarying(program, index, maxLen - 1, length, size, type, name);
+    gl.glGetTransformFeedbackVarying(program.id, index, maxLen - 1, length, size, type, name);
     checkError('getTransformFeedbackVarying');
     int _type = type.value;
     String _name = name.cast<Utf8>().toDartString();
@@ -409,9 +409,16 @@ class RenderingContext {
   /// Be careful which type of integer you really pass here. Unfortunately an UInt16List
   /// is viewed by the Dart type system just as List<int>, so we jave to specify the native type
   /// here in [nativeType]
-  void bufferData(int target, NativeArray data, int usage) {
+  void bufferData(int target, dynamic data, int? usage) {
     startCheck('bufferData');
-    gl.glBufferData(target, data.lengthInBytes, data.data, usage);
+    if(data is int){
+      var offSetPointer = Pointer<Void>.fromAddress(data);
+      gl.glBufferData(target, data, offSetPointer.cast(), usage ?? 0);
+    }
+    else{
+      gl.glBufferData(target, data.lengthInBytes, data.data, usage ?? 0);
+    }
+    
     checkError('bufferData');
   }
 
@@ -464,7 +471,7 @@ class RenderingContext {
           final infoLog = calloc<Int8>(infoLen.value);
 
           gl.glGetShaderInfoLog(shader.id, infoLen.value, nullptr, infoLog);
-          message = "\nError compiling shader:\n${infoLog.cast<Utf8>().toDartString()}";
+          message = "\nError compiling shader:\n${infoLog.cast<Utf8>()}";
 
           calloc.free(infoLog);
         }
@@ -496,8 +503,14 @@ class RenderingContext {
   // void copyTexSubImage2D(int target, int level, int xoffset, int yoffset, int x, int y, int width, int height);
   void copyTexSubImage2D(int target, int level, int xoffset, int yoffset, int x, int y, int width, int height){
     startCheck('copyTexSubImage2D');
-    gl.glCopyTexSubImage2D(target, level, xoffset, yoffset, x,y,width, height);
+    gl.glCopyTexSubImage2D(target, level, xoffset, yoffset, x, y, width, height);
     checkError('copyTexSubImage2D');
+  }
+
+  void copyTexSubImage3D(int target, int level, int xoffset, int yoffset, int zoffset, int x, int y, int width, int height){
+    startCheck('copyTexSubImage3D');
+    gl.glCopyTexSubImage3D(target, level, xoffset, yoffset, zoffset, x, y, width, height);
+    checkError('copyTexSubImage3D');
   }
 
   Buffer createBuffer() {
@@ -676,7 +689,7 @@ class RenderingContext {
     var offSetPointer = Pointer<Void>.fromAddress(offset);
     gl.glDrawElements(mode, count, type, offSetPointer.cast());
     checkError('drawElements');
-    calloc.free(offSetPointer);
+    //calloc.free(offSetPointer);
   }
 
   void enable(int cap) {
@@ -738,7 +751,8 @@ class RenderingContext {
     checkError('getActiveAttrib');
 
     int _type = type.value;
-    String _name = name.cast<Utf8>().toDartString();
+    Pointer<Utf8> temp = name.cast<Utf8>();
+    String _name = temp == nullptr?'unnamed':temp.toDartString();
     int _size = size.value;
 
     calloc.free(type);
@@ -760,7 +774,8 @@ class RenderingContext {
     checkError('getActiveUniform');
 
     int _type = type.value;
-    String _name = name.cast<Utf8>().toDartString();
+    Pointer<Utf8> temp = name.cast<Utf8>();
+    String _name = temp == nullptr?'unnamed':temp.toDartString();
     int _size = size.value;
 
     calloc.free(type);
@@ -795,11 +810,11 @@ class RenderingContext {
     if (Platform.isMacOS) {
       return getExtensionMacos(key);
     }
-    Pointer _v = gl.glGetString(WebGL.EXTENSIONS);
+    Pointer<Uint8> _v = gl.glGetString(WebGL.EXTENSIONS);
     checkError('getExtension');
-    String _vstr = _v.cast<Utf8>().toDartString();
+    String _vstr = _v == nullptr?'unnamed':_v.cast<Utf8>().toDartString();
     List<String> _extensions = _vstr.split(" ");
-    // calloc.free(_v);
+    //calloc.free(_v);
     return _extensions;
   }
 
@@ -816,11 +831,11 @@ class RenderingContext {
 
   String getStringi(int key, int index) {
     startCheck('getStringi');
-    Pointer _v = gl.glGetStringi(key, index);
+    Pointer<Uint8> _v = gl.glGetStringi(key, index);
     checkError('getStringi');
-    String temp = _v.cast<Utf8>().toDartString();
+    String str = _v == nullptr?'unnamed':_v.cast<Utf8>().toDartString();
     //calloc.free(_v);
-    return temp;
+    return str;
   }
 
   int getIntegerv(int v0) {
@@ -849,7 +864,7 @@ class RenderingContext {
       final infoLog = calloc<Int8>(_len);
       gl.glGetProgramInfoLog(program.id, _len, nullptr, infoLog);
       checkError('getProgramInfoLog');
-      message = "\nError compiling shader:\n${infoLog.cast<Utf8>().toDartString()}";
+      message = "\nError compiling shader:\n${infoLog.cast<Utf8>()}";
       calloc.free(infoLog);
       return message;
     } 
@@ -882,7 +897,7 @@ class RenderingContext {
 
       gl.glGetShaderInfoLog(shader.id, _len, nullptr, infoLog);
       checkError('getShaderInfoLog');
-      message = "\nError compiling shader:\n${infoLog.cast<Utf8>().toDartString()}";
+      message = "\nError compiling shader:\n${infoLog.cast<Utf8>()}";
       calloc.free(infoLog);
       return message;
     }
@@ -898,9 +913,9 @@ class RenderingContext {
     calloc.free(_pointer);
     return _v == 0?false:true;
   }
-  void framebufferTextureLayer(int target,int attachment,int texture,int level,int layer){
+  void framebufferTextureLayer(int target,int attachment,WebGLTexture? texture,int level,int layer){
     startCheck('framebufferTextureLayer');
-    gl.glFramebufferTextureLayer(target, attachment, texture, level, layer);
+    gl.glFramebufferTextureLayer(target, attachment, texture?.id, level, layer);
     checkError('framebufferTextureLayer');
   }
 
@@ -958,7 +973,7 @@ class RenderingContext {
           final infoLog = calloc<Int8>(infoLen.value);
 
           gl.glGetProgramInfoLog(program.id, infoLen.value, nullptr, infoLog);
-          message = "\nError linking program:\n${infoLog.cast<Utf8>().toDartString()}";
+          message = "\nError linking program:\n${infoLog.cast<Utf8>()}";
 
           calloc.free(infoLog);
         }
@@ -1117,7 +1132,7 @@ class RenderingContext {
 
   void uniform1fv(UniformLocation location, List<double> v){
     startCheck('uniform1fv');
-    var arrayPointer = floatListToArrayPointer(v);
+    var arrayPointer = _floatListToArrayPointer(v);
     gl.glUniform1fv(location.id, v.length, arrayPointer);
     checkError('uniform1fv');
     calloc.free(arrayPointer);
@@ -1146,7 +1161,7 @@ class RenderingContext {
 
   void uniform2fv(UniformLocation location, List<double> v){
     startCheck('uniform2fv');
-    var arrayPointer = floatListToArrayPointer(v);
+    var arrayPointer = _floatListToArrayPointer(v);
     gl.glUniform2fv(location.id, v.length ~/ 2, arrayPointer);
     checkError('uniform2fv');
     calloc.free(arrayPointer);
@@ -1174,7 +1189,7 @@ class RenderingContext {
 
   void uniform3fv(UniformLocation location, List<double> vectors) {
     startCheck('uniform3fv');
-    var arrayPointer = floatListToArrayPointer(vectors);
+    var arrayPointer = _floatListToArrayPointer(vectors);
     gl.glUniform3fv(location.id, vectors.length ~/ 3, arrayPointer);
     checkError('uniform3fv');
     calloc.free(arrayPointer);
@@ -1197,7 +1212,7 @@ class RenderingContext {
 
   void uniform4fv(UniformLocation location, List<double> vectors) {
     startCheck('uniform4fv');
-    var arrayPointer = floatListToArrayPointer(vectors);
+    var arrayPointer = _floatListToArrayPointer(vectors);
     gl.glUniform4fv(location.id, vectors.length ~/ 4, arrayPointer);
     checkError('uniform4fv');
     calloc.free(arrayPointer);
@@ -1214,7 +1229,7 @@ class RenderingContext {
 
   void uniformMatrix2fv(UniformLocation location, bool transpose, List<double> values) {
     startCheck('uniformMatrix2fv');
-    var arrayPointer = floatListToArrayPointer(values);
+    var arrayPointer = _floatListToArrayPointer(values);
     gl.glUniformMatrix2fv(location.id, values.length ~/ 4, transpose ? 1 : 0, arrayPointer);
     checkError('uniformMatrix2fv');
     calloc.free(arrayPointer);
@@ -1222,7 +1237,7 @@ class RenderingContext {
 
   void uniformMatrix3fv(UniformLocation location, bool transpose, List<double> values) {
     startCheck('uniformMatrix3fv');
-    var arrayPointer = floatListToArrayPointer(values);
+    var arrayPointer = _floatListToArrayPointer(values);
     gl.glUniformMatrix3fv(location.id, values.length ~/ 9, transpose ? 1 : 0, arrayPointer);
     checkError('uniformMatrix3fv');
     calloc.free(arrayPointer);
@@ -1231,7 +1246,7 @@ class RenderingContext {
   /// be careful, data always has a length that is a multiple of 16
   void uniformMatrix4fv(UniformLocation location, bool transpose, List<double> values) {
     startCheck('uniformMatrix4fv');
-    var arrayPointer = floatListToArrayPointer(values);
+    var arrayPointer = _floatListToArrayPointer(values);
     gl.glUniformMatrix4fv(location.id, values.length ~/ 16, transpose ? 1 : 0, arrayPointer);
     checkError('uniformMatrix4fv');
     calloc.free(arrayPointer);
@@ -1245,7 +1260,7 @@ class RenderingContext {
 
   void vertexAttrib1fv(int index, List<double> values){
     startCheck('vertexAttrib1fv');
-    var arrayPointer = floatListToArrayPointer(values);
+    var arrayPointer = _floatListToArrayPointer(values);
     gl.glVertexAttrib1fv(index, arrayPointer);
     checkError('vertexAttrib2fv');
     calloc.free(arrayPointer);
@@ -1253,7 +1268,7 @@ class RenderingContext {
 
   void vertexAttrib2fv(int index, List<double> values){
     startCheck('vertexAttrib2fv');
-    var arrayPointer = floatListToArrayPointer(values);
+    var arrayPointer = _floatListToArrayPointer(values);
     gl.glVertexAttrib2fv(index, arrayPointer);
     checkError('vertexAttrib2fv');
     calloc.free(arrayPointer);
@@ -1261,7 +1276,7 @@ class RenderingContext {
 
   void vertexAttrib3fv(int index, List<double> values){
     startCheck('vertexAttrib3fv');
-    var arrayPointer = floatListToArrayPointer(values);
+    var arrayPointer = _floatListToArrayPointer(values);
     gl.glVertexAttrib3fv(index, arrayPointer);
     checkError('vertexAttrib3fv');
     calloc.free(arrayPointer);
@@ -1269,7 +1284,7 @@ class RenderingContext {
 
   void vertexAttrib4fv(int index, List<double> values){
     startCheck('vertexAttrib4fv');
-    var arrayPointer = floatListToArrayPointer(values);
+    var arrayPointer = _floatListToArrayPointer(values);
     gl.glVertexAttrib4fv(index, arrayPointer);
     checkError('vertexAttrib4fv');
     calloc.free(arrayPointer);
@@ -1309,9 +1324,9 @@ class RenderingContext {
 
   void vertexAttribPointer(int index, int size, int type, bool normalized, int stride, int offset) {
     startCheck('vertexAttribPointer');
-    var offsetPointer = Pointer<Void>.fromAddress(offset);
+    Pointer<Void> offsetPointer = Pointer<Void>.fromAddress(offset);
     using((Arena arena) {
-      gl.glVertexAttribPointer(index, size, type, normalized ? 1 : 0, stride, offsetPointer.cast<Void>());
+      gl.glVertexAttribPointer(index, size, type, normalized ? 1 : 0, stride, offsetPointer);
     });
     checkError('vertexAttribPointer');
     //calloc.free(offsetPointer);
@@ -1329,7 +1344,7 @@ class RenderingContext {
     checkError('readPixels');
   }
 
-  Pointer<Float> floatListToArrayPointer(List<double> list) {
+  Pointer<Float> _floatListToArrayPointer(List<double> list) {
     final ptr = calloc<Float>(list.length);
     ptr.asTypedList(list.length).setAll(0, list);
     return ptr;