+++ flutter_angle/lib/webgl/angle.dart
@@ -1,40 +1,42 @@
 import '../shared/classes.dart';
 import '../shared/options.dart';
 import 'dart:async';
-import 'dart:html';
+import 'dart:js_interop';
+import 'package:web/web.dart' as html;
 import 'wrapper.dart';
 import 'gles_bindings.dart';
 import 'dart:ui_web' as ui;
 import 'dart:math' as math;
 
 class FlutterAngleTexture {
-  final CanvasElement? element;
+  final html.HTMLCanvasElement? element;
   final int textureId;
   final int rboId;
-  final int metalAsGLTextureId;
-  late final int androidSurface;
+  final int surfaceId;
   final int fboId;
+  final int loc;
   LibOpenGLES? _libOpenGLES;
   late AngleOptions options;
   
   FlutterAngleTexture(
+    FlutterAngle flutterAngle,
     this.textureId, 
     this.rboId, 
-    this.metalAsGLTextureId,
-    this.androidSurface, 
+    this.surfaceId,
     this.element,
-    this.fboId, 
+    this.fboId,
+    this.loc,
     this.options
-  );
+  ) {}
 
   LibOpenGLES get rawOpenGl {
     if (_libOpenGLES == null) {
       _libOpenGLES = LibOpenGLES(
-        element?.getContext(
+         element?.getContext(
           "webgl2", {
             "alpha": options.alpha, 
             "antialias": options.antialias
-          }
+          }.jsify()
         )!
       );
     }
@@ -42,28 +44,10 @@ class FlutterAngleTexture {
     return _libOpenGLES!;
   }
 
-  static FlutterAngleTexture fromMap(
-    dynamic map, 
-    CanvasElement? element,
-    int fboId, 
-    AngleOptions options
-  ) {
-    return FlutterAngleTexture(
-      map['textureId']! as int,
-      map['rbo'] as int? ?? 0,
-      map['metalAsGLTexture'] as int? ?? 0,
-      map['surface'] as int? ?? 0,
-      element,
-      fboId,
-      options
-    );
-  }
-
   Map<String, int> toMap() {
     return {
       'textureId': textureId,
       'rbo': rboId,
-      'metalAsGLTexture': metalAsGLTextureId
     };
   }
 
@@ -75,20 +59,13 @@ class FlutterAngleTexture {
   /// As you can have multiple Texture objects, but WebGL allways draws in the currently
   /// active one you have to call this function if you use more than one Textureobject before
   /// you can start rendering on it. If you forget it you will render into the wrong Texture.
-  void activate() {
-    //rawOpenGl.glViewport(0, 0, width, height);
-  }
-
+  void activate() {}
   RenderingContext getContext() {
     return RenderingContext.create(rawOpenGl,options.width, options.height);
   }
 }
 
 class FlutterAngle{
-
-  static Future<String> get platformVersion async {
-    return 'WebGL';
-  }
 
   static void glDebugOutput(
     int source, 
@@ -100,29 +77,35 @@ class FlutterAngle{
     int pUserParam
   ) {}
 
-  static Future<FlutterAngleTexture> createTexture(AngleOptions options) async {
+  Future<FlutterAngleTexture> createTexture(AngleOptions options) async {
     final _divId = DateTime.now().microsecondsSinceEpoch;
-    final element = CanvasElement(
-      width: (options.width * options.dpr).toInt(), 
-      height: (options.height * options.dpr).toInt()
-    )..id = 'canvas-id${math.Random().nextInt(100)}';
+    final element = html.HTMLCanvasElement()
+    ..width = (options.width * options.dpr).toInt()
+    ..height = (options.height * options.dpr).toInt()
+    ..id = 'canvas-id${math.Random().nextInt(100)}';
 
     ui.platformViewRegistry.registerViewFactory(_divId.toString(), (int viewId) {
       return element;
     });
 
-    final newTexture = FlutterAngleTexture.fromMap({
-      "textureId": _divId
-    }, element, 0, options);
+    final newTexture = FlutterAngleTexture(
+      this,
+      _divId,
+      0,0,
+      element, 
+      0,0,
+      options
+    );
 
     return newTexture;
   }
 
-  static Future<void> initOpenGL([bool useDebugContext = false, bool useAngle = false]) async {}
-  static Future<void> updateTexture(FlutterAngleTexture texture,[WebGLTexture? sourceTexture]) async {
+  Future<void> init([bool useDebugContext = false, bool useAngle = false]) async {}
+  Future<void> updateTexture(FlutterAngleTexture texture,[WebGLTexture? sourceTexture]) async {
     texture.rawOpenGl.glFlush();
   }
-  static Future<void> deleteTexture(FlutterAngleTexture texture) async {}
-  static void activateTexture(FlutterAngleTexture texture) {}
-  static void printOpenGLError(String message) {}
+  Future<void> deleteTexture(FlutterAngleTexture texture) async {}
+  void activateTexture(FlutterAngleTexture texture) {}
+  void printOpenGLError(String message) {}
+  void dispose([List<FlutterAngleTexture>? textures]){}
 }