+++ flutter_angle_native_array/lib/flutter_angle_native_array.dart
@@ -1,5 +1,0 @@
-library flutter_angle_native_array;
-
-export 'native_array_app.dart' if (dart.library.js) './native_array_web.dart';
-export 'base_native_array.dart';
-export 'native_array.dart';